drm/amd/dal: temporary fix in headless processing code for MST
We cannot rely in function aconnector_from_drm_crtc_id for crtc_id to
be equal to connector_id. It is not the case anymore. Because of that
there was a segfault:

1. Pageflip atomic commit;
2. No connector provided in state, so aconnector is NULL;
3. Incorrect physical MST connector selected in handle_headless_hotplug,
and set mode is scheduled for it;
4. Segfault because there is no modes exist on this connector.

We need to revisit handle_headless_hotplug logic, and retest them.

Signed-off-by: Harry Wentland <harry.wentland@amd.com>
Acked-by: Harry Wentland <harry.wentland@amd.com>